requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-smart-selects==1.5.9
djangorestframework==3.12.1
Markdown==3.3.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
